test: :white_check_mark: add tests for delete and get all measures

+++ src/modules/use-cases/delete-measure/delete-measure.use-case.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { IMeasureRepository } from '../../measure/repository/measure.interface.repository';
 
 @Injectable()
@@ -7,6 +7,16 @@ export class DeleteMeasureUseCase {
   private readonly measureRepository: IMeasureRepository;
 
   async execute(measure_uuid: string) {
+    const measure = await this.measureRepository.findMeasureById(measure_uuid);
+
+    if (measure === undefined) {
+      throw new NotFoundException({
+        error_code: 'MEASURE_NOT_FOUND',
+        error_description: 'Leitura não encontrada',
+      });
+    }
     await this.measureRepository.deleteMeasure(measure_uuid);
+
+    return true;
   }
 }
